chore: move schema to shared
