+++ src/node-lib/curriculum-api-2023/queries/browseUnitRedirect/browseUnitRedirect.query.ts
@@ -1,7 +1,4 @@
-import {
-  redirectSchema,
-  Redirect,
-} from "../canonicalLessonRedirect/canonicalLessonRedirect.schema";
+import { redirectSchema, Redirect } from "../../shared.schema";
 
 import errorReporter from "@/common-lib/error-reporter";
 import OakError from "@/errors/OakError";
